Use ngram  tokenizer instead of ngram filter.
Desire was to remove punctuation characters, such as hypen and quote.

The most straightforward path was to use the keyword tokenizer,
and then use the word_delimiter filter to remove those characters.

The issue with this approach is the whole word (part numbers in this
case) are highlighted, not just the portion that match.